# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BSD-YBM OS Workspace API
  version: 1.0.0
  description: תת-קבוצה של endpoints ליבה — הרחבה הדרגתית ל-10/10 הנדסי.
servers:
  - url: https://www.bsd-ybm.co.il
    description: Production
  - url: http://localhost:3000
    description: Local dev
paths:
...
  /api/crm/contacts:
    get:
      summary: List CRM contacts with ERP summary
      security:
        - sessionCookie: []
    post:
      summary: Create CRM contact
      security:
        - sessionCookie: []
...
  /api/erp/archive/bulk-export:
    post:
      summary: Bulk export archive items as ZIP (max 50)
      security:
        - sessionCookie: []
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token